+++ aws-resources/src/main/java/io/opentelemetry/contrib/aws/resource/internal/AwsResourceDetector.java
@@ -3,9 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.aws.resource;
+package io.opentelemetry.contrib.aws.resource.internal;
 
 import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
+import io.opentelemetry.contrib.aws.resource.BeanstalkResource;
+import io.opentelemetry.contrib.aws.resource.Ec2Resource;
+import io.opentelemetry.contrib.aws.resource.EcsResource;
+import io.opentelemetry.contrib.aws.resource.EksResource;
+import io.opentelemetry.contrib.aws.resource.LambdaResource;
 import io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider;
 import io.opentelemetry.sdk.resources.Resource;
 import io.opentelemetry.sdk.resources.ResourceBuilder;
