+++ aws-resources/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider
@@ -1,1 +1,1 @@
-io.opentelemetry.contrib.aws.resource.AwsResourceDetector
+io.opentelemetry.contrib.aws.resource.internal.AwsResourceDetector
