+++ gcp-resources/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider
@@ -1,1 +1,1 @@
-io.opentelemetry.contrib.gcp.resource.GcpResourceDetector
+io.opentelemetry.contrib.gcp.resource.internal.GcpResourceDetector
